Fixes to build on Linux

+++ src/main.cpp
@@ -18,7 +18,7 @@
 int main()
 {
     uint32_t nEpochs = 5;
-    uint32_t miniBatchSize = 10;
+    uint16_t miniBatchSize = 10;
     float learningRate = 3.0f;
     std::vector<uint16_t> layersizes{784, 30, 10};
